src/test/java/com/codurance/training/tasks/usecase/DeleteTaskUseCaseTest.java
package com.codurance.training.tasks.usecase;

import com.codurance.training.tasks.adapter.port.out.InMemoryProjectRepository;
import com.codurance.training.tasks.adapter.port.out.ProjectStore;
import com.codurance.training.tasks.adapter.port.out.TaskStore;
import com.codurance.training.tasks.framework.persistant.ProjectMemoryStore;
import com.codurance.training.tasks.framework.persistant.TaskMemoryStore;
import com.codurance.training.tasks.usecase.port.in.project.add.AddProjectUseCase;
import com.codurance.training.tasks.usecase.port.in.task.add.AddTaskUseCase;
import com.codurance.training.tasks.usecase.port.in.task.delete.DeleteTaskUseCase;
import com.codurance.training.tasks.usecase.port.out.repository.ProjectRepository;
import org.junit.Before;
import org.junit.Test;

import java.util.HashMap;

import static org.junit.Assert.*;

public class DeleteTaskUseCaseTest {
    private ProjectRepository repository;
    @Before
    public void setUp() {
        ProjectStore projectStore = new ProjectMemoryStore(new HashMap<>());
        TaskStore taskStore = new TaskMemoryStore(new HashMap<>());
        repository = new InMemoryProjectRepository(projectStore, taskStore);
    }

    @Test
    public void add_task_in_project_test() {
        AddProjectUseCase addProjectUseCase = new AddProjectUseCase(repository);
        addProjectUseCase.execute("test");
        AddTaskUseCase addTaskUseCase = new AddTaskUseCase(repository);
        addTaskUseCase.execute("test", "First task", "T01");

        DeleteTaskUseCase deleteTaskUseCase = new DeleteTaskUseCase(repository);
        deleteTaskUseCase.execute("T01");

        assertFalse(repository.findTask("T01").isPresent());
        assertEquals(0, repository.findProject("test").get().getProjectTasks().size());
    }
}
